# review: perf: override AbstractMap#get in ElementNameMap