@I-Al-Istannen 2022-08-19
I crunched some numbers. The initial regression from @slarse's PR was approximately a factor of 2 (not placing blame here by any means, I didn't see it either and the change was good).

I ran my javadoc indexer for various configurations and got:

### Before slarse order fix
`266s` (user) and `36` wall clock time

### After slarse order fix
`683s` (user) and `1:23` wall clock time

### After this PR
`440s` (user) and `55s` wall clock time

### After this PR and hannes/perf/PackageFactory
`249s` (user) and `35s` wall clock time


Both changes from @SirYwell address unfortunate consequences of the inherited `AbstractMap` methods I didn't foresee.
@SirYwell and I really need to come up with some continuous benchmark integration, it seems. I didn't expect the impact of the PR to be that high and should have probably measured it.